chore(cqrs): split files to follow single-type-per-file convention
